@@ -17,24 +17,21 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        Toolbar toolbar =  findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
+        FloatingActionButton fab =  findViewById(R.id.fab);
+        fab.setOnClickListener(view -> {
+            Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
+                    .setAction("Action", null).show();
 
-                ULog.v("onClick");
-                ULog.i("onClick");
-                ULog.d("onClick");
-                ULog.w("onClick");
-                ULog.e("onClick");
-                ULog.wtf("onClick");
-                ULog.wtf("------------------------");
-            }
+            ULog.v("onClick");
+            ULog.i("onClick");
+            ULog.d("onClick");
+            ULog.w("onClick");
+            ULog.e("onClick");
+            ULog.wtf("onClick");
+            ULog.wtf("------------------------");
         });
 
         ULog.v("onCreate");
